cluster benchmarks, removed some code duplication
