+++ heat/utils/data/__init__.py
@@ -6,3 +6,4 @@ from .datatools import *
 from . import matrixgallery
 from . import mnist
 from .partial_dataset import *
+from . import spherical
